bcachefs: avoid wrapped shift in inline data cut_back
Same class as the fiemap fix: k.k->size is u32 sectors, and
k.k->size << 9 wraps in 32 bits for sizes >= 1 << 23. Legitimate
inline_data keys can't get there (the val is bounded by bkey val
capacity), but a corrupted key with a huge size would wrap the shift
before min() clamps it, and bch2_cut_back_s() would then truncate the
val to the wrong size - compounding the corruption instead of handling
it. Widen before shifting.

Found by an audit of all 130 sector-to-byte shift sites in fs/,
prompted by the fiemap truncation fix; this was the only other site
where the wrap is reachable (and only via a corrupt key).

+++ fs/data/extents.c
@@ -2083,7 +2083,7 @@ int bch2_cut_back_s(struct bpos where, struct bkey_s k)
 	case KEY_TYPE_inline_data:
 	case KEY_TYPE_indirect_inline_data:
 		new_val_u64s = (bkey_inline_data_offset(k.k) +
-				min(bkey_inline_data_bytes(k.k), k.k->size << 9)) >> 3;
+				min(bkey_inline_data_bytes(k.k), (u64) k.k->size << 9)) >> 3;
 		break;
 	}
